Refactor/Separate useAssetContractGetter from useBorrowSubmit
